# backend/ferko-web-api/src/main/java/hr/fer/zemris/ferko/webapi/dto/CreateToDoTaskRequest.java
package hr.fer.zemris.ferko.webapi.dto;

import java.time.LocalDateTime;

public record CreateToDoTaskRequest(
    long assigneeId, String title, String description, LocalDateTime deadline, String priority) {}